+++ crates/nostr/src/nips/nip01.rs
@@ -17,8 +17,8 @@ use serde::ser::{SerializeMap, Serializer};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use super::nip19::FromBech32;
-use super::nip21::FromNostrUri;
+use super::nip19::{self, FromBech32, Nip19Coordinate, ToBech32};
+use super::nip21::{FromNostrUri, ToNostrUri};
 use crate::types::Url;
 use crate::{key, Filter, JsonUtil, Kind, PublicKey, Tag};
 
@@ -222,6 +222,27 @@ impl FromStr for Coordinate {
     }
 }
 
+impl ToBech32 for Coordinate {
+    type Err = nip19::Error;
+
+    #[inline]
+    fn to_bech32(&self) -> Result<String, Self::Err> {
+        self.borrow().to_bech32()
+    }
+}
+
+impl FromBech32 for Coordinate {
+    type Err = nip19::Error;
+
+    fn from_bech32(addr: &str) -> Result<Self, Self::Err> {
+        let coordinate: Nip19Coordinate = Nip19Coordinate::from_bech32(addr)?;
+        Ok(coordinate.coordinate)
+    }
+}
+
+impl ToNostrUri for Coordinate {}
+impl FromNostrUri for Coordinate {}
+
 /// Borrowed coordinate
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct CoordinateBorrow<'a> {
@@ -245,6 +266,17 @@ impl CoordinateBorrow<'_> {
         }
     }
 }
+
+impl ToBech32 for CoordinateBorrow<'_> {
+    type Err = nip19::Error;
+
+    #[inline]
+    fn to_bech32(&self) -> Result<String, Self::Err> {
+        nip19::coordinate_to_bech32(*self, &[])
+    }
+}
+
+impl ToNostrUri for CoordinateBorrow<'_> {}
 
 /// Metadata
 #[derive(Debug, Clone, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]